docs(components/atom/label): label documentation articles
